Added featured videos below main vid on watch page Extracted fetchRecommendation function to separate file for reusability Extracted types to d.ts files for reusability #8
