+++ app/app/src/pages/Watch.tsx
@@ -7,6 +7,8 @@ import DownloadButton from "../components/DownloadButton";
 import {Theme} from "@material-ui/core";
 import Toast from "../components/Toast";
 import {getUrl} from "../utils"
+import FeaturedVideos from "../components/FeaturedVideos";
+import {getRecommendations} from "../services/recommendations";
 
 function Watch() {
     const [searchParams] = useSearchParams();
@@ -15,6 +17,7 @@ function Watch() {
     const [muted, setMuted] = useState(false)
     const [image, setImage] = useState()
     const [downloading, setDownloading] = useState(false)
+    const [featuredVideos, setFeaturedVideos] = useState<any[]>([])
 
     const hashCode = (input: string) => {
         var hash = 0, i, chr;
@@ -54,42 +57,20 @@ function Watch() {
         window.URL.revokeObjectURL(url);
     }
 
-    const handleSubmitImage = async (e: File) => {
-        // setShowPlayer(true)
-
-        const arrBuffer = await e.arrayBuffer()
-        if (arrBuffer === undefined) {
-            return
-        }
-
-        const base64String = window.btoa(new Uint8Array(arrBuffer).reduce(function (data, byte) {
-            return data + String.fromCharCode(byte);
-        }, ''));
-        const resp = await fetch(getUrl() + "/img", {
-            method: "POST",
-            body: JSON.stringify({img: base64String, width: window.innerWidth / 4})
-        })
-        if (resp.status !== 200) {
-            console.error("error uploading image:", await resp.text())
-            alert("error processing image")
-            // setShowPlayer(false)
-        }
-
-        const decoded = await resp.json()
-        console.info(decoded)
-        setImage(decoded)
-        setPlayerContent(decoded.img)
-        // setShowPlayer(true)
-    }
-
     useEffect(() => {
         setMuted(true)
+    }, [])
+
+    useEffect(() => {
+        getRecommendations().then(recommendations => {
+            setFeaturedVideos(recommendations)
+        })
     }, [])
 
     return (
         <>
             <Grid container>
-                <Grid item xs={9}>
+                <Grid item xs={12}>
                     <AsciiPlayer
                         ytURL={`https://www.youtube.com/watch?v=${searchParams.get('v')}`}
                         image={image}
@@ -107,10 +88,8 @@ function Watch() {
                             </div>
                     }
                 </Grid>
-                <Grid item xs={3}>
-                    video list
-                </Grid>
             </Grid>
+            <FeaturedVideos videos={featuredVideos} />
             {muted ? <Toast text={(isMobile ? "tap" : "click") + " to unmute"}/> : null}
         </>
     )
