+++ app/app/src/types/FeaturedVideos.d.ts
@@ -1,0 +1,8 @@
+import YoutubeVideoThumbnailProps from "./YoutubeVideoThumbnail";
+
+export default interface FeaturedVideosProps {
+    videos: {
+        id: number,
+        props: YoutubeVideoThumbnailProps,
+    }[]
+}
