+++ app/app/src/types/YoutubeVideoThumbnail.d.ts
@@ -1,0 +1,15 @@
+export default interface YoutubeVideoThumbnailProps {
+    youtubeId: string
+    title: string
+    thumbnailImage: {
+        src: string
+    }
+    avatarImage: {
+        alt: string
+        src: string
+    }
+    channel: {
+        youtubeId: string
+        title: string
+    }
+}
